+++ RmbCloneDb/dbo/Tables/Location.sql
@@ -2,6 +2,6 @@
 (
 	[Id] NVARCHAR(128) NOT NULL PRIMARY KEY, 
     [Address] NVARCHAR(50) NOT NULL, 
-    [Latitude] REAL NOT NULL, 
-    [Longitude] REAL NOT NULL
+    [Latitude] DECIMAL(10, 7) NOT NULL, 
+    [Longitude] DECIMAL(10, 7) NOT NULL
 )
